Generic controller: fix classID setting in ClassElement.
The code previously stored the (probably unmodified) id
of its wrapped source class only when registration *failed*,
that is an error.

As shown by `git grep -B2 setID` the correct behaviour that
all other elements follow is to update their ID when registration
succeeded.

The dubious question of why this information has to be duplicated in
the first place is out of scope for now :)