Generic controller: fix REG_CONTROLLER
As there are no tests for this plugin, I can only assume that
registration of sources and sinks by the controller plugin has
never actually been tested, or at least not against the
current version of the audio manager.

This behaviour is triggered by using REG_CONTROLLER as
the registration type when declaring a sink or a source.

The problem is as follows:

* The elements and element classes are created from the xml file

* As no element class has been registered yet, the
  class ID of the newly-created elements is initially
  set to 1 (in CAmConfigurationReader.cpp)

* The element classes are registered, they now have a valid class ID

* The elements are then registered, which fails because core checks
  that the class ID is valid in CAmDatabaseHandlerMap::enterSourceDB
  and CAmDatabaseHandlerMap::enterSinkDB

The proposed solution is to update the class ID of sinks and
sources along with their domainID at registration time.