PluginControlGeneric: fix SourceSetVolume logic.
It is possible that the action was invoked only with a
volume parameter, or only with a limit parameter. In
either case, the other parameter should not be considered
to have a default set, as with for example the volume, which
was set to 0 when only the limit parameter was set.