ALSA RA: Initial release of ALSA Routing Adapter plugin for AudioManager.
The ALSA Routing Adapter allows to manage ALSA applications.
Applications have to interact with the AudioManager command interface.
The registration of sink, sources and gateways is done with an XML file.
You can configure so called proxies to stream between different ALSA devices on request.
The volume control interface is implemented and requires the softvol plugin or
hardware which supports volume control via simple alsa controls.

Signed-off-by: Jens Lorenz <jlorenz@de.adit-jv.com>